+++ packages/remax/src/adapters/wechat/components/CoverView.ts
@@ -8,7 +8,7 @@ export interface CoverViewProps {
   className?: string;
   style?: CSSProperties;
   scrollTop?: number | string; //		否	设置顶部滚动偏移量，仅在设置了 overflow-y: scroll 成为滚动元素后生效	2.1.0
-  onClick?: (event: any) => void;
+  onClick?: (event: any) => any;
   animation?: Record<string, any>[];
 }