+++ packages/remax/src/adapters/wechat/components/MovableArea.ts
@@ -8,7 +8,7 @@ export interface MovableAreaProps {
   className?: string;
   style?: CSSProperties;
   scaleArea?: boolean; // false 否 当里面的movable-view设置为支持双指缩放时，设置此值可将缩放手势生效区域修改为整个movable-area 1.9.90
-  onClick?: (event: any) => void;
+  onClick?: (event: any) => any;
 }
 
 export default MovableArea;