+++ packages/remax/src/adapters/wechat/components/Picker.ts
@@ -9,8 +9,8 @@ export interface PickerProps {
   style?: CSSProperties;
   mode?: string; // selector 否 选择器类型 1.0.0
   disabled?: boolean; // false 否 是否禁用 1.0.0
-  onClick?: (event: any) => void;
-  onCancel?: (event: any) => void; //  否 取消选择时触发 1.9.90
+  onClick?: (event: any) => any;
+  onCancel?: (event: any) => any; //  否 取消选择时触发 1.9.90
   animation?: Record<string, any>[];
 }